requirements.txt
numpy
torch
torchvision
transformers
fastapi
Pillow
uvicorn[standard]
kornia
einops
timm
python-multipart
pydantic
ollama
opencv-python
openai
python-dotenv
loadimg
markitdown[all]